moved logic out of chat button to be exclusively in the new component
