+++ packages/react-composites/src/composites/MeetingComposite/ChatButton.tsx
@@ -4,33 +4,15 @@
 import React from 'react';
 import { ControlBarButton, ControlBarButtonProps } from '@internal/react-components';
 import { Chat20Filled, Chat20Regular } from '@fluentui/react-icons';
-import { NotificationIcon } from './NotificationIcon';
-import { IStackStyles, Stack } from '@fluentui/react';
-import { ChatButtonContainerStyles } from './styles/MeetingCompositeStyles';
+
+const onRenderOnIcon = (): JSX.Element => <Chat20Filled key={'chatOnIconKey'} primaryFill="currentColor" />;
+const onRenderOffIcon = (): JSX.Element => <Chat20Regular key={'chatOffIconKey'} primaryFill="currentColor" />;
 
 /**
  * @private
  */
-export interface ChatButtonProps extends ControlBarButtonProps {
-  chatButtonChecked: boolean | undefined;
-  unreadMessageCount: number;
-}
-/**
- * @private
- */
-export const ChatButton = (props: ChatButtonProps): JSX.Element => {
+export const ChatButton = (props: ControlBarButtonProps): JSX.Element => {
   const strings = { label: props.label, ...props.strings };
-  const { unreadMessageCount } = props;
-
-  const onRenderOnIcon = (): JSX.Element => <Chat20Filled key={'chatOnIconKey'} primaryFill="currentColor" />;
-  const onRenderOffIcon = (): JSX.Element => {
-    return (
-      <Stack styles={chatNotificationContainerStyles}>
-        {unreadMessageCount > 0 && <NotificationIcon chatMessagesCount={unreadMessageCount} />}
-        <Chat20Regular key={'chatOffIconKey'} primaryFill="currentColor" />
-      </Stack>
-    );
-  };
 
   return (
     <ControlBarButton
@@ -42,11 +24,4 @@ export const ChatButton = (props: ChatButtonProps): JSX.Element => {
       onClick={props.onClick}
     />
   );
-};
-
-const chatNotificationContainerStyles: IStackStyles = {
-  root: {
-    display: 'inline',
-    position: 'relative'
-  }
 };
